Update tests for postgres 9.4 9.5
