+++ tests/test_default.py
@@ -18,22 +18,12 @@ def test_databases(Sudo, Command, name, expected_db):
     assert out == expected_db
 
 
-@pytest.mark.parametrize("name,expected_roles", [
-    ('alice', 'alice|f|t|f|f|f|t|f|-1|********|||'),
-    ('bob', 'bob|f|t|f|t|f|t|f|-1|********|||'),
-])
-def test_user_roles(Sudo, Command, name, expected_roles):
-    sql = "SELECT * FROM pg_roles WHERE rolname='%s'" % name
-    with Sudo('postgres'):
-        out = Command.check_output('psql postgres -c "%s" -At' % sql)
-    # Everything except the UID at the end
-    assert out.startswith(expected_roles)
-
-
 def test_server_listen(File, Sudo, TestinfraBackend):
     host = TestinfraBackend.get_hostname()
+    ver = match('postgresql-(\d)(\d)-\w+', host).group(1, 2)
     with Sudo():
-        f = File('/var/lib/pgsql/9.4/data/postgresql.conf').content_string
+        f = File('/var/lib/pgsql/%s.%s/data/postgresql.conf' % ver
+                 ).content_string
 
     count_listen_addresses = 0
     for line in f.split('\n'):
@@ -43,6 +33,6 @@ def test_server_listen(File, Sudo, TestinfraBackend):
     assert count_listen_addresses == 1
 
     if host == 'postgresql-94-all':
-        assert listen_addresses == "listen_addresses = '\*'"
+        assert listen_addresses == "listen_addresses = '*'"
     else:
         assert listen_addresses == "listen_addresses = localhost"
